Sidebar layout and contents
Layout the areas of the sidebar and put the contents in for each of them; an icon (img) and a string (p). There are three areas; the title line which holds the dashboard icon and string, the navigation links, and the other links.

Note: In the code before I had written the sidebar as "dashboard" for classes, so a good chunk of the code has been renamed to use "sidebar" instead of "dashboard" for better clarity.

index.html
- Add all the code for the sidebar layout and contents
 - This is a lot of nested divs

files
- Added an images folder and added all the images for the sidebar

styles.css
- The sidebar is a grid with 3 children
 - The title line is a grid, while the other children have their own .sidebar-line children which are grids