Format the main content container so cards can go in it
The container is a grid itself, and then the projects and announcements sections are also grids. So far the work has only been done on the Your Projects section as I wanted to work on one of them first. The cards for this section have also had their initial styling so I could check to see if the grid was working properly.

files
- Added a content folder to the images folder to hold the icons each project card uses

index.html
- Added the code for several cards as well as headers for the project and announcement (other-cards) sections
 - A card is itself a grid which contains 2ps (a heading and content) and a final grid which holds the icons for doing stuff for that card/project

styles.css
- Added all the CSS needed for the project container
 - Most notable is that the project container is a grid, each card is a grid, and inside each grid is a card links grid
 - Almost the entirety of the CSS is defining grids, but there are a couple things for color and font stuff where needed
- Cards are made of 3 sub-classes which fit into its grid
 - A heading
 - Content
 - A card links section, which is also a grid